requirements.txt
addict==2.4.0
contourpy==1.3.0
cycler==0.12.1
fonttools==4.60.2
ftfy==6.3.1
h5py==3.14.0
ImageIO==2.37.2
importlib_resources==6.5.2
kiwisolver==1.4.7
markdown-it-py==3.0.0
matplotlib==3.9.4
mdurl==0.1.2
mmcv==2.0.0
mmengine==0.10.7
-e git+https://github.com/open-mmlab/mmsegmentation.git@b040e147adfa027bbc071b624bedf0ae84dfc922#egg=mmsegmentation
numpy==1.26.4
opencv-python==4.8.1.78
packaging==26.0
pillow==11.3.0
platformdirs==4.4.0
prettytable==3.16.0
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2026.1.15
rich==14.3.2
scipy==1.13.1
six==1.17.0
termcolor==3.1.0
tomli==2.4.0
torch==1.10.2+cu111
torchvision==0.11.3+cu111
tqdm==4.67.3
typing_extensions==4.15.0
wcwidth==0.6.0
yapf==0.43.0
zipp==3.23.0
